chore(mobile): Defaults WebView route capture simplifying WebView usage throughout
