+++ apps/mobile/src/components/HyloWebView/HyloWebView.js
@@ -1,11 +1,12 @@
 import React, { useCallback, useState, useEffect } from 'react'
-import { useFocusEffect } from '@react-navigation/native'
+import { useFocusEffect, useNavigation } from '@react-navigation/native'
 import Config from 'react-native-config'
-import Loading from 'components/Loading'
 import useRouteParams from 'hooks/useRouteParams'
 import AutoHeightWebView from 'react-native-autoheight-webview'
 import queryString from 'query-string'
 import { WebViewMessageTypes } from '@hylo/shared'
+import useOpenURL from 'hooks/useOpenURL'
+import { modalScreenName } from 'hooks/useIsModalScreen'
 import { getSessionCookie } from 'util/session'
 import { match, pathToRegexp } from 'path-to-regexp'
 import { parseWebViewMessage } from '.'
@@ -36,24 +37,62 @@ const handledWebRoutesJavascriptCreator = loadedPath => allowRoutesParam => {
   `
 }
 
+const useNativeRouteHandler = () => {
+  const navigation = useNavigation()
+  const openURL = useOpenURL()
+
+  return ({ pathname, search }) => ({
+    '(.*)/:type(post|members)/:id': ({ routeParams }) => {
+      const { type, id } = routeParams
+
+      switch (type) {
+        case 'post': {
+          navigation.navigate('Post Details', { id })
+          break
+        }
+        case 'members': {
+          navigation.navigate('Member', { id })
+          break
+        }
+      }
+    },
+    '(.*)/post/:postId/edit': ({ routeParams }) => {
+      navigation.navigate('Edit Post', { id: routeParams.postId })
+    },
+    '(.*)/group/:groupSlug([a-zA-Z0-9-]+)': ({ routeParams }) => {
+      navigation.navigate(modalScreenName('Group Explore'), routeParams)
+    },
+    '/:groupSlug(all)/topics/:topicName': ({ routeParams: { topicName } }) => {
+      navigation.navigate('Stream', { topicName })
+    },
+    '(.*)/topics/:topicName': ({ routeParams: { topicName } }) => {
+      navigation.navigate('Chat Room', { topicName })
+    },
+    '(.*)': () => {
+      openURL(pathname + search)
+    }
+  })
+}
+
 const HyloWebView = React.forwardRef(({
   handledWebRoutes = [],
   messageHandler,
-  nativeRouteHandler,
+  nativeRouteHandler: nativeRouteHandlerProp,
   path: pathProp,
-  route,
   style,
   source,
   ...forwardedProps
 }, webViewRef) => {
   const [cookie, setCookie] = useState()
   const [uri, setUri] = useState()
-  const { postId, path: routePath } = useRouteParams()
+  const { postId, path: routePath, originalLinkingPath } = useRouteParams()
+  const nativeRouteHandler = nativeRouteHandlerProp || useNativeRouteHandler()
+
+  const path = pathProp || routePath || originalLinkingPath || ''
 
   useEffect(() => {
-    const path = pathProp || routePath
-    setUri((source?.uri || `${Config.HYLO_WEB_BASE_URL}${path || ''}`) + (postId ? `?postId=${postId}` : ''))
-  }, [source?.uri, pathProp, routePath])
+    setUri((source?.uri || `${Config.HYLO_WEB_BASE_URL}${path}`) + (postId ? `?postId=${postId}` : ''))
+  }, [source?.uri, pathProp, routePath, originalLinkingPath])
 
   useFocusEffect(
     useCallback(() => {
@@ -99,13 +138,13 @@ const HyloWebView = React.forwardRef(({
     messageHandler && messageHandler(parsedMessage)
   }
 
-  if (!cookie) return <Loading />
+  if (!cookie || !uri) return null
 
   return (
     <AutoHeightWebView
       customScript={`
         window.HyloWebView = true;
-        ${pathProp && handledWebRoutesJavascriptCreator(pathProp)(handledWebRoutes)}
+        ${path && handledWebRoutesJavascriptCreator(path)(handledWebRoutes)}
       `}
       geolocationEnabled
       onMessage={handleMessage}
