+++ apps/mobile/src/navigation/HomeNavigator.js
@@ -14,6 +14,7 @@ import AllViews from 'screens/AllViews'
 import Groups from 'screens/Groups'
 import GroupWelcome from 'screens/GroupWelcome'
 import GroupWelcomePage from 'screens/GroupWelcome/GroupWelcomePage'
+import HyloWebView from 'components/HyloWebView'
 import MemberDetails from 'screens/MemberProfile/MemberDetails'
 import MemberProfile from 'screens/MemberProfile'
 import MembersComponent from 'screens/Members'
@@ -59,6 +60,7 @@ export default function HomeNavigator () {
       <HomeTab.Screen name='Chat Room' component={ChatRoomWebView} />
       <HomeTab.Screen name='Group Settings' component={GroupSettingsWebView} />
       <HomeTab.Screen name='User Settings' component={UserSettingsWebView} />
+      <HomeTab.Screen name='Web View' component={HyloWebView} />
       {/* Other screens */}
       <HomeTab.Screen name='Stream' component={Stream} />
       <HomeTab.Screen name='All Views' component={AllViews} />
