+++ apps/mobile/src/screens/GroupExploreWebView/GroupExploreWebView.js
@@ -7,7 +7,7 @@ import { DEFAULT_APP_HOST } from 'navigation/linking'
 import groupDetailsQueryMaker from '@hylo/graphql/queries/groupDetailsQueryMaker'
 import useGroup from '@hylo/hooks/useGroup'
 import { openURL } from 'hooks/useOpenURL'
-import useIsModalScreen, { modalScreenName } from 'hooks/useIsModalScreen'
+import useIsModalScreen from 'hooks/useIsModalScreen'
 import useRouteParams from 'hooks/useRouteParams'
 import ModalHeaderTransparent from 'navigation/headers/ModalHeaderTransparent'
 import HyloWebView from 'components/HyloWebView'
@@ -77,29 +77,6 @@ export default function GroupExploreWebView () {
     '(.*)/explore/group/(.*)'
   ]
 
-  const nativeRouteHandler = ({ pathname, search }) => ({
-    '(.*)/:type(post|members)/:id': ({ routeParams }) => {
-      const { type, id } = routeParams
-
-      switch (type) {
-        case 'post': {
-          navigation.navigate('Post Details', { id })
-          break
-        }
-        case 'members': {
-          navigation.navigate('Member', { id })
-          break
-        }
-      }
-    },
-    '/groups/:groupSlug': ({ routeParams }) => {
-      navigation.navigate(modalScreenName('Group Explore'), routeParams)
-    },
-    '(.*)': () => {
-      openURL(pathname + search)
-    }
-  })
-
   if (!groupSlug) return null
 
   return (
@@ -107,7 +84,6 @@ export default function GroupExploreWebView () {
       ref={webViewRef}
       path={path}
       handledWebRoutes={handledWebRoutes}
-      nativeRouteHandler={nativeRouteHandler}
       messageHandler={messageHandler}
       // TODO: Consider adding this to the `HyloWebView` standard API
       onNavigationStateChange={({ url, canGoBack: providedCanGoBack }) => {
